feat: Working on the second week's projects
-Completing the hangman project. This project only needs error handling
-Weather forcast app: I have coded the GUI interface with one label and
a text field to input the name of the country plus a button to execute the
program, the GUI has been tested successfully. I've also written the code
for the API calls to get the forcast data, I've stored the values of the
weather status and the icons in a list, this part wasn't tested ( because I'm
missing an API key :p ) but it should work fine.